framework-comparisons/02-multi-agent-orchestration/google-adk/requirements.txt
google-adk==1.22.1
python-dotenv==1.2.1
